Update create post form
