+++ misc/markdown.js
@@ -2,42 +2,8 @@
 // This takes input through STDIN and spits HTML to STDOUT. Please cache all
 //  things in redis
 
-var MarkdownIt = require('markdown-it'),
-    KaTeX = require('markdown-it-katex'),
-    hljs = require('highlight.js'),
-    footnote = require('markdown-it-footnote'),
-    mila = require('markdown-it-link-attributes'),
-    os = require('os');
-
-let md = new MarkdownIt({
-    html: false, // No html = no xss
-    linkify: true,
-    highlight: function (string, language) {
-        var langExec = "";
-        if (language) {
-            langExec = ' exec-target" data-lang="' + language;
-        }
-        if (language && hljs.getLanguage(language)) {
-            try {
-                return '<pre class="hljs' + langExec + '"><code>' +
-                    hljs.highlight(language, string, true).value +
-                    '</code></pre>';
-            } catch (__) {
-                return '<div style="font-size: 3rem; color: red; font-weight: 300">Error Rendering</div>';
-            }
-        }
-
-        return '<pre class="hljs' + langExec + '"><code>' + md.utils.escapeHtml(string) + '</code></pre>';
-    }
-}).use(KaTeX, {
-    'throwOnError': false,
-    'errorColor': '#cc0000'
-}).use(mila, {
-    attrs: {
-        target: '_blank',
-        rel: 'nofollow'
-    }
-}).use(footnote);
+var md = require('./markdown-renderer'),
+    os = require('os')
 
 process.stdin.resume();
 process.stdin.setEncoding('utf-8');
