+++ package.json
@@ -15,10 +15,11 @@
     "envify": "^4.1.0",
     "escape-html": "^1.0.3",
     "highlight.js": "^9.12.0",
+    "katex": "^0.9.0-beta1",
     "markdown-it": "^8.4.0",
     "markdown-it-footnote": "^3.0.1",
-    "markdown-it-katex": "^2.0.3",
     "markdown-it-link-attributes": "^2.0.0",
+    "markdown-it-texmath": "^0.4.7",
     "n": "^2.1.8",
     "progress": "^2.0.0",
     "uglify-js": "^3.3.4"
