+++ static/scss/body.scss
@@ -45,8 +45,6 @@ a {
 }
 
 blockquote {
-
-
   margin: 0;
   padding: .7rem 2rem;
   padding-right: 0;
@@ -55,12 +53,14 @@ blockquote {
   border-left: 8px solid $blockquote-marker;
   border-radius: 3px;
 
+  color: rgba($blockquote-fg, 0.7);
+
   p {
     font-size: 1.3rem;
     font-family: $quote-font;
     font-style: italic;
 
-    color: rgba($blockquote-fg, 0.7);
+    color: inherit;
 
     &:last-child { margin-bottom: 0; }
   }
@@ -78,6 +78,12 @@ li, p {
   // Default margin results in no extra space.
   // This adds a margin-top if there is an element after katex
   margin-top: 1em;
+
+  // Slightly bigger (as display)
+  font-size: 17px;
+
+  // Makes KaTeX correct color
+  color: inherit;
 }
 
 table {
