+++ static/scss/main.scss
@@ -62,16 +62,17 @@ body {
 
         display: flex;
         flex-direction: row;
-        align-items: stretch;
 
         > main {
-          align-self: start;
+          align-self: stretch;
 
           width: 0px; // Prevents expansion of parent
           flex: 1;
 
           > form {
-            flex: 1;
+            // at least full width.
+            min-height: 100%;
+
             display: flex;
             flex-direction: column;
           }
