+++ static/scss/modal.scss
@@ -24,9 +24,9 @@
 
     overflow-y: auto;
 
-    box-shadow: 0 0 8px #888;
+    box-shadow: 0 0 50px rgba(0, 0, 0, 0.4);
 
-    background: white;
+    background: $dim-bg;
     border-radius: 16px;
     padding: 0em 1em;
     box-sizing: border-box;
