+++ static/scss/nodes.scss
@@ -111,5 +111,7 @@ hr {
 
 .body {
   overflow: auto;
+  color: $main-text;
+
   @import "body.scss";
 }
