+++ templates/layouts/page.html
@@ -14,7 +14,7 @@
     <link type="text/css" href="https://fonts.googleapis.com/css?family=Lato:900|Roboto:300,400,700" rel="stylesheet">
     <link type="text/css" href="https://fonts.googleapis.com/css?family=Fira+Mono&amp;subset=greek,latin-ext"
           rel="stylesheet">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/KaTeX/0.6.0/katex.min.css"
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/KaTeX/0.8.3/katex.min.css"
           crossorigin="anonymous">
 
     {% if g.user is none -%}
